Export minimal with new columns
