+++ packages/react-components/lib/tasks/index.ts
@@ -5,3 +5,4 @@ export * from './task-table';
 export * from './task-timeline';
 export * from './task-table-datagrid';
 export * from './task-schedule-event-edit-delete-popup';
+export * from './utils';
